# services/order/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Order API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3004"
    }
  ],
  "tags": [
    {
      "name": "Order",
      "description": "API for managing Orders"
    },
    {
      "name": "OrderItem",
      "description": "API for managing Order Items"
    },
    {
      "name": "Tracking",
      "description": "API for tracking Order statuses"
    },
    {
      "name": "Product",
      "description": "API for managing Products"
    }
  ],
  "paths": {
    "/orders": {
      "get": {
        "summary": "Get all orders",
        "tags": ["Order"],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new order and send to RabbitMQ",
        "description": "This endpoint creates a new order, saves it to the database, and sends the order details to a RabbitMQ queue for processing.",
        "tags": ["Order"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created and message sent to RabbitMQ"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get an order by ID",
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an order's status",
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated"
          }
        }
      },
      "delete": {
        "summary": "Delete an order",
        "tags": ["Order"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted"
          }
        }
      }
    },
    "/orders/{id}/items": {
      "get": {
        "summary": "Get all order items by order ID",
        "tags": ["OrderItem"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of order items",
            "content": {
              "application/json": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OrderItem"
                }
              }
            }
          }
        }
      }
    },
    "/tracking/{order_id}": {
      "get": {
        "summary": "Get tracking info for an order",
        "tags": ["Tracking"],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Tracking details",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/Tracking"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update tracking info",
        "tags": ["Tracking"],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTracking"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tracking updated"
          }
        }
      },
      "delete": {
        "summary": "Delete tracking info",
        "tags": ["Tracking"],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Tracking deleted"
          }
        }
      }
    },
    "/tracking": {
      "get": {
        "summary": "Get all tracking",
        "tags": ["Tracking"],
        "responses": {
          "200": {
            "description": "List of tracking",
            "content": {
              "application/json": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tracking"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create tracking for an order",
        "tags": ["Tracking"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTracking"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tracking created"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "tags": ["Product"],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "tags": ["Product"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get product details by ID",
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update product details",
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "order_id": { "type": "string" },
          "customer_id": { "type": "string" },
          "total_amount": { "type": "number" },
          "status": { "type": "string" },
          "order_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "NewOrder": {
        "type": "object",
        "properties": {
          "customer_id": { "type": "string" },
          "status": { "type": "string" },
          "order_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "product_id": { "type": "string" },
          "quantity": { "type": "integer" },
          "price_per_unit": { "type": "number" }
        }
      },
      "UpdateOrderStatus": {
        "type": "object",
        "properties": {
          "status": { "type": "string" }
        }
      },
      "Tracking": {
        "type": "object",
        "properties": {
          "order_id": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "tracking_status": { "type": "string" }
        }
      },
      "NewTracking": {
        "type": "object",
        "properties": {
          "order_id": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "tracking_status": { "type": "string" }
        }
      },
      "UpdateTracking": {
        "type": "object",
        "properties": {
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "tracking_status": { "type": "string" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "product_id": { "type": "string" },
          "product_name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" }
        }
      },
      "NewProduct": {
        "type": "object",
        "properties": {
          "product_name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" }
        }
      },
      "UpdateProduct": {
        "type": "object",
        "properties": {
          "product_name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" }
        }
      }
    }
  }
}